Moves logger into base and factory into different namespace
