+++ lib/alephant/logger.rb
@@ -1,33 +1,19 @@
-require 'alephant/logger/version'
-require 'logger'
+require "alephant/logger/version"
+require "alephant/logger_factory"
+require "logger"
 
 module Alephant
   module Logger
-    def self.create(drivers = [])
-      Logger.new drivers
+    @@logger = Alephant::LoggerFactory.create
+
+    def logger
+      @@logger
     end
 
-    class Logger
-      def initialize(drivers)
-        @drivers = drivers << ::Logger.new(STDOUT)
-      end
-
-      def method_missing(name, *args)
-        drivers.each do |driver|
-          driver.send(name, *args) if driver.respond_to? name
-        end
-      end
-
-      def respond_to?(name)
-        drivers.any? do |driver|
-          driver.respond_to?(name) || super
-        end
-      end
-
-      private
-
-      attr_reader :drivers
+    def self.setup(*drivers)
+      @@logger = Alephant::LoggerFactory.create(drivers.flatten)
     end
+
   end
 end
